debugged message route

+++ app.py
@@ -10,15 +10,11 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_debugtoolbar import DebugToolbarExtension
 
 # Local modules
-# import urllib
 import helper
 import model
 
 
 app = flask.Flask(__name__)
-
-# Required to use Flask sessions. - REMOVE
-# app.secret_key = "ABC"
 
 
 @app.errorhandler(400)
@@ -75,7 +71,7 @@ def sign_up():
     return response
 
 
-# curl -k -H "Content-Type: application/json" -X POST -d '{"username":"foo","password":"bar"}' http://localhost:5000/log_in
+# curl -k -H "Content-Type: application/json" -X POST -d '{"username":"foo","password":"bar"}' -c cookie-jar.txt http://localhost:5000/log_in
 @app.route('/log_in', methods=['POST'])
 def log_in():
     """Endpoint for logging in.
@@ -108,7 +104,7 @@ def log_in():
     return response
 
 
-# curl -k -i -H "Accept: application/json" -H "Content-Type: application/json" -X GET http://localhost:5000/log_out
+# curl -k -i -H "Accept: application/json" -H "Content-Type: application/json" -X GET -b cookie-jar.txt http://localhost:5000/log_out
 @app.route('/log_out', methods=['GET'])
 def log_out():
     """Endpoint for logging out.
@@ -116,15 +112,19 @@ def log_out():
     Response:
         {'status': <status>, 'msg': <msg>}"""
 
+    # Only log out if user logged in.
+    if not flask.request.cookies.get('session_id'):
+        return bad_request()
+
     model.Session.query.filter(model.Session.session_id
                                == flask.request.cookies.get('session_id')).delete()
-    import pdb; pdb.set_trace()
     response = flask.jsonify ({'status' : 'OK',
                                'msg' : 'User successfully logged out'})
     return response
 
 
-# curl -k -H "Content-Type: application/json" -X POST -d '{"sender_id":"1","recipient_id":"2","message_content":"hello world"}' http://localhost:5000/messages
+# curl -k -X POST -d '{"sender_id":"1","recipient_id":"2","message_content":"hello world"}' -b cookie-jar.txt http://localhost:5000/messages
+# curl -k -H "Content-Type: application/json" -X POST -d '{"sender_id":"1","recipient_id":"2","message_content":"hello world"}' -b cookie-jar.txt http://localhost:5000/messages
 @app.route('/messages', methods=['POST'])
 def send_message():
     """Endpoint to accept a sender, recipient and message and insert them into the data store.
@@ -138,49 +138,52 @@ def send_message():
     Response:   
         {'status': <status>, 'msg': <msg>}"""
 
-    import pdb; pdb.set_trace()
     message = json.loads(flask.request.data)
     session = model.Session.query.filter(model.Session.session_id
                                          == flask.request.cookies.get('session_id')).first()
 
     # Only accept message if a token was passed with the request
     # and the token returns the correct user ID.
-    if session.user_id != message['sender_id']:
+    if session.user_id != int(message['sender_id']):
         return bad_request()
 
     conversation_id = helper.make_conversation_id(message['sender_id'],
                                                   message['recipient_id'])
     timestamp = datetime.datetime.utcnow()
+
     if message['message_content'].startswith('http'):
 
         # For purposes of the challenge, meta data is harcoded
         # In practice, image width, height are included in magic.from_buffer return value
         if helper.is_image(message['message_content']):
-            message_type = 'image_link'
-            message_metadata = flask.jsonify({'img_width': '300',  # Image width is stored in px
-                                              'img_height': '500'})  # Image height is stored in px
+            message_type = model.MessageType.query.filter(model.MessageType.message_type 
+                                                        == 'image_link').first()
+            message_metadata = {'img_width': '300',  # Image width is stored in px
+                                'img_height': '500'}  # Image height is stored in px
             
         # For purposes of the challenge, meta data is harcoded
         # In practice, consider extracting domain name from HTTP header
         # & vid_length from APIs for whitelisted video platforms
         elif helper.is_video(message['message_content']):
-            message_type = 'video_link'
-            message_metadata = flask.jsonify({'vid_length': '180',  # Video length is stored in seconds
-                                              'vid_source': 'youtube'})
-
-    elif flask.request.headers['Content-Type'] == 'text/plain':
-        message_type = 'text'
-        messages_metadata = flask.jsonify({})
+            message_type = model.MessageType.query.filter(model.MessageType.message_type 
+                                                        == 'video_link').first()
+            message_metadata = {'vid_length': '180',  # Video length is stored in seconds
+                                'vid_source': 'youtube'}
+        else:
+            return unsupported_media_type()
 
     else:
-        return unsupported_media_type()
+        message_type = model.MessageType.query.filter(model.MessageType.message_type 
+                                                    == 'text').first()
+        message_metadata = {}
 
-    new_message = modelMessage(conversation_id=conversation_id,
+    message_id = message_type.message_type_id
+    new_message = model.Message(conversation_id=conversation_id,
                                sender_id=message['sender_id'],
-                               message_type=message_type,
+                               message_type=message_id,
                                message_content=message['message_content'],
                                timestamp=timestamp,
-                               message_metadata=message_metadata)
+                               message_metadata=json.dumps(message_metadata))
     model.db.session.add(new_message)
     model.db.session.commit()
     response = flask.jsonify({'status' : 'OK',
@@ -201,6 +204,11 @@ def fetch_messages():
         {'status': <status>, 'msg': <msg>}"""
 
     fetch = json.loads(flask.request.data)
+    
+    # Only return messages if user logged in.
+    if not flask.request.cookies.get('session_id'):
+        return bad_request()
+
     conversation_id = helper.make_conversation_id(fetch['user_id_1'],
                                                   fetch['user_id_2'])
     # Sort by descending time.
